fix(header): more robust header styles
buttons should really only be 2ems. 2.5 looks a little silly.
However, this was making the header too thin, so the solution
was to put wrappers around the header cells and middle-align
the contents. Using <label> as the wrapper makes it so you can
click anywhere in the cell and the button will get clicked,
even though the button doesn't visually take up the whole height
of the header.

Finally, the triple-dot "more" menu wasn't displaying inline.
It should now show up at the right side of the header, as it was
meant to.