fix(login): Change "Sign up" to register to reduce confusion
"Sign in" and "Sign up" look very similar and could cause users
to pause to think about which one. Even though both paths will
ultimately do the same thing, we don't want users to think about
which button to push. It should just be clear.

Note that the reason we have two distinct buttons in the first
place is to avoid the situation where users only see one button
that they assume doesn't apply to them.